+++ Content.Server/Imperial/_SystemsSupplementation/Zombies/Components/ZombieBarotraumaDamageComponent.cs
@@ -1,7 +1,7 @@
 using Content.Shared.Damage;
 using Content.Shared.FixedPoint;
 
-namespace Content.Shared.Imperial.Zombies;
+namespace Content.Server.Zombies;
 
 [RegisterComponent]
 public sealed partial class ZombieBarotraumaDamageComponent : Component
@@ -10,7 +10,7 @@ public sealed partial class ZombieBarotraumaDamageComponent : Component
     /// Компонент, хранящий информацию о том,
     /// какой урон будет наносится сущности после зомбификации от компонента BarotraumaComponent.
     /// </summary>
-    [DataField("damage")]
+    [DataField]
     public DamageSpecifier Damage = new()
     {
         DamageDict = new() { { "Blunt", FixedPoint2.New(0.20) } }
